requirements.txt
numpy
pandas
Pillow
plotly
streamlit
tensorflow
Keras
opencv-python